requirements.txt
customtkinter
pytesseract
pillow
python-docx
pypdf
tkinterdnd2
